requirements.txt
# Trading & Market Data
alpaca-py>=0.43.0,<1.0
pytz>=2024.0

# LLM (google-generativeai is deprecated since Nov 2025)
google-genai>=1.60.0,<2.0

# Configuration
pyyaml>=6.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0

# Environment
python-dotenv>=1.0

# Data
pandas>=2.0

# HTTP & retry
tenacity>=8.0
httpx>=0.25

# Logging
structlog>=24.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=4.0
freezegun>=1.2
